+++ frontend/techpick/src/hooks/useSearchElementId.ts
@@ -1,9 +1,8 @@
-import { useEffect, useState } from 'react';
 import { useSearchParams } from 'next/navigation';
+import { useEffect, useState } from 'react';
 
 export default function useSearchElementId() {
   const searchElementId = useSearchParams().get('searchId');
-  const dateId = useSearchParams().get('dateId');
   const [isOccurClickEvent, setIsOccerClickEvent] = useState<boolean>(false);
 
   useEffect(() => {
@@ -27,7 +26,7 @@ export default function useSearchElementId() {
       document.removeEventListener('click', handleClick);
       setIsOccerClickEvent(false);
     };
-  }, [searchElementId, dateId]);
+  }, [searchElementId]);
 
   return {
     searchElementId,
